Rename *DEPEND to *REQUIRES, add BUILD_PREREQUIRES
The renaming aligns the nomenclature with that of the package infos. The new
BUILD_PREREQUIRES lists build prerequisites that cannot be built first. E.g.
gcc requires gcc for building, and make which in turn must be built with gcc.
The union of all ports' BUILD_PREREQUIRES is the set of packages needed for
bootstrapping the repository build. Those are expected to be installed in the
host system already. Eventually we will want to set up a cross-compilation
environment for those, but that isn't an immediate concern.