Evaluate PATCHES variable set in recipes
Allows recipes to explicitly specify the patches to apply.

If a patch file with the name of the port exists, PATCHES is set to
that by default. This feature should rather be removed eventually, so
that all patches have to be specified explicitly. I was just too lazy
to adjust all concerned recipes.