Add some more configure directory variables
The added ones seem to be accepted even by older versions of autotools
generated build systems, so that hopefully doesn't break any previously
building package. libExecDir is mapped to "bin/private",
{shared,local}StateDir to "settings".

Also add a comment with the missing directories (newer build systems),
which have defaults that should be OK for us for now.