Implement updating the final package requires
Add class RequiresUpdater. It is initialized with the packages required
for building the port. It extracts their provides declaration and
later, when creating the final package info for a package, updates the
versions in the requires list according to the matching provides.

May not work correctly when cross-building. E.g. for required commands
we might update the version to the build host's command version.