Fix package info overwrite problem
If the file to be written exists Package._generatePackageInfo() does
now remove it first. In case of a symlink it would otherwise overwrite
the linked filed. This happened whenever building a port. The port's
package infos in the (global) repository would be overwritten with
files not containing any provides declaration.