BuildPlatformUnix: simplify include/lib dir setup
ATM the include and lib directories we create for the cross tools are
there mostly for aesthetical reasons. ATM they are just symlinks to the
respective directories of the extracted Haiku sysroot package and are
used as the target of the symlinks we create in the actual cross tools
directory. So we can just as well use a simple directory layout.