/boot/common has been removed
Mostly the changes are straight-forward (don't look in /boot/common,
adjust paths). A notable change concerns the resolution of port
dependencies: We no longer simply include all /boot/system packages in
the chroot, but only those that have to be pulled in according to the
dependency resolution. This may break some recipes that don't declare
all the things they actually require for building, but have worked
previously due to those dependencies living in /boot/system.