Allow a port to be built if only the secondary arch is marked as supported.
Part of #66. This removes the need for the somewhat-cryptic if-blocks
that check $effectiveTargetArchitecture for compatibility. It does break
recipes that need to have declared support for both primary & secondary
architectures, but AFAIK none of those exist (yet), so when they do, we
can add another case for them (or let them use the if-syntax).