Add option --no-system-packages to constrain used packages.
If supplied the system packages directory will not be added to the
list of package repositories used to resolve dependencies.

This means that only local packages from the packages path will be used
for the search of all dependency packages, including the search for the
main Haiku package itself. It therefore requires one to provide all
needed packages as part of the local packages but allows for more
control over what exact packages are visible.