BuildMaster: Implement concept of builder vs. build failure.
When a build fails due to a builder issue (IO errors, connection issues
or failure to connect in the first place), reschedule the build so it
can be built on another builder.

Builders are then marked as either unavailable (which will cause further
builds to be scheduled and a new connection attempted) or lost. Lost
builders aren't added to the list of available builders anymore.