BuildMaster: Detect builder disconnect differently.
Apparently no socket.error is raised when the builder disconnects
while a command is running. Instead the command just returns with a
-1 exit code meaning "no exit code provided by server". Since negative
numbers aren't valid exit codes, check if the connection is still
alive when a negative exit code is returned and mark the builder as
unavailable when it disconnected.