BuildMaster: Ensure a consistent schedule before running.
Ensure that the schedule can actually work, i.e. all packages that
blocked builds are missing will actually be built as part of the run.

The schedule can be inconsistent due to bugs elsewhere that yet need to
be fixed. For now just mark the builds lost that won't be satisfied by
the run. Once the issues are fixed this should be turned into an assert.

Also keep a list of complete, failed and lost builds.