Configuration: handle default repositoryDirectory earlier.
When REPOSITORY_DIRECTORY is not set, its default value was
being set in Repository.py. This might be a bad idea because
Configuration.getRepositoryDirectory() is not aware of that
default value.
The change in this commit update moves to Configuration.py
the definition of repositoryDirectory in case there was no
REPOSITORY_DIRECTORY definition in haikuports.conf.
To make things simple, the suffix that was previously being
added to the default repository directory is no longer used.

New commit summary in case this commit update is accepted:

Configuration: new optional REPOSITORY_DIRECTORY parameter

to allow relocation/specification of the repository directory.

While at it, drop the suffix ('-' + $targetArchitecture) that
was added to the default path for the repository directory, in
the previous commit. The default path is therefore restored:
$OUTPUT_DIRECTORY/repository

    Of course, this change is neutral if you don't set that parameter.