BuildMaster: Rework the idle builder mechanism.
Make the idle/active builder lists in the status simply depend on
whether a builder is currently doing a build.

This partially reverts 400e9470ecf63bf0c67930bbe50244677b9fe01e which
actually moved the idle builders at the end of a buildrun into a
separate list. This caused a regression when a builder failed in that
phase, as its build could not be rescheduled since there were no
builders in the active list anymore.