PackageRepository: Add class to manage the packages directory.
Initially it allows pruning packages, which does:

* Obsolete packages that are newer than the active port version, with
the assumption that the active version was downgraded.
* Obsolete packages that have newer version packages available.

This is slightly different from normal package obsoletion in that it
does keep older versions of packages as long as no newer ones are
available.

The use case for this is to prepare a packages directory to be usable
for repo creation and continuous updates.