ProvidesManager: Ignore HPKGs when 'base' is required.
In the general package management sense, the 'base' modifier ensures
that a package is installed in the same location as its base package.

Within HaikuPorter it also declares the relationship between a sub
package of a port and its base package. A base package requirement
should therefore always be resolved within that port and can't come
from a HPKG.