BuildMaster: Skip builds of ports that already have packages.
This strictly requires the version to be changed (by bumping the
revision for example) so that no packages are produced that would
replace existing ones.

Packages must generally be considered immutable, as the package manager
has no way to detect the need to update a package apart from its
version. No changes may therefore be made to existing packages that
might already have been served. Since there is always an implicit
requires update when building packages with HaikuPorter, even seemingly
unproblematic changes may cause the package to be altered in a relevant
way.