Main: Exit buildmaster when instructed to build obsolete ports.
This can happen when dependencies require specific versions that are
not active.

For example this may happen when already present packages were built
against a version of a port that is later deactivated. In such a case,
these packages need to be rebuilt to not depend on the unavailable
port anymore. They would otherwise be uninstallable in any case.