PackageInfo: Fix resolvable compat parsing for recipe format.
The manual parsing of the compatibility version did only take the format
used in HPKGs into account, i.e. "(compatible >= version)". It is
however also used on the expressions from recipe files, which use the
format "compat >= version". This lead to the compat version being
ignored and therefore possibly wrong dependencies to be used.

Remove the manual parsing and put it into the regular expression. Also
simplify the logic by using optional groups in the regular expression
so that the match result can be used regardless of what parts exist in
the string.