DependencyResolver: Try harder to resolve dependencies.
When a dependency of a package can't be met, remove the provides of that
package from the ProvidesManager and restart the dependency resolution.

This allows to still find valid solutions where dependencies can be met
by multiple sources of which earlier ones fail to have their requires
fulfilled.

For example:

program.recipe requires lib:first
  libfirst-1.0.hpkg provides lib:first requires lib:second >= 1
  libfirst-1.1.recipe provides lib:first requires lib:second >= 2
    libsecond-2.0.hpkg provides lib:second == 2 compat >= 2

In the case above, the dependency lib:second of libfirst-1.0.hpkg
can't be resolved, because libsecond-2.0.hpkg provides an incompatible
version. However, the recipe for libfirst-1.1 would build a compatible
package and should therefore be used instead. Before this change,
dependency resolution would fail at the former. Now it will remove
the source of the provides that can't have its dependencies met,
libfirst-1.0.hpkg in the example above, and restart the dependency
resolution so that libfirst-1.1.recipe will be picked up and built.

The dependency resolution failures are collected and a combined message
including all of them is presented if the dependency resolution
ultimately still fails. This has the nice effect of showing the whole
dependency tree that failed, not just the last leave, which should make
it more obvious why a certain, possibly indirect, unresolved dependency
was pulled in in the first place.