Repository: Reverse logic of stale port removal.
Only ports with broken recipes had their .DependencyInfo and port
for package mapping removed. This missed various cases of how a port
could otherwise become unavailable, for example by getting deactivated
or by having a newer version becoming active.

Instead of building a list of ports to remove, build a list of active
ports and remove what isn't on it. This results in the same repository
state as a fresh population would result in.

This fixes the case where .DependencyInfos of deactivated ports would
linger on, breaking dependency resolution in non-obvious ways.