Repository: Remove superfluous higherVersionIsActive logic.
The goal is to update the repository only with the active versions. As
the version list is sorted from highest to lowest, the first port that
fully runs through is the active one and anything else can be skipped.

The only reason to set higherVersionIsActive and continue the loop was
to remove the dependency infos of the next lower version. This assumed
that old versions would stay in the tree as opposed to being removed
completely. It also only cared for the dependency info and missed the
removal from the port for package mappings.

Since there already is a stale port removal mechanism that takes all
needed steps into account, this duplicate logic can simply be dropped.