Repository: Drop/add port for package mappings of updated ports.
Previously the mappings were simply added. This made it possible for
removed sub-packages to remain in the mappings, since stale removal
only works on a per port basis and the port remained active.

Drop the mappings for updated ports by initially not adding updated
ports to the activePorts list when running stale mappings removal. Then
add the fresh mappings.

This also critically moves the removal of stale dependency infos past
the removal of stale port for package mappings. This is needed because
the former uses the latter to determine if dependency infos should
be kept. With the inverse order used before, dependency infos for
removed packages would only be removed in the next repository update.