Port: Rework configuration of repository dir.
Use a class variable and a setter to make the repository dir known to
Port and add a context manager for temporarily adjusting the repository
dir of specific ports.

This removes the need to hand in the repository dir to various methods,
which makes things more consistent. Some methods were previously using
the repository dir given to the constructor while others needed it as
argument.

The Repository now makes itself known to Port and can change the current
repository dir for the purpose of populating a new repository at a
different path.

With this, the ugly manual adjustment of the recipe cache of the Port by
the Repository can be removed. It was directly changing internals of
Port but used slightly different logic than within Port which lead to
two different filenames being used for the recipe cache file. This
caused recipes to be parsed once unnecessarily.