Repository/Main: Rework handling of obsolete port building.
Repository: The method supportBackwardsCompatibility was a near 100%
duplicate of the original _updateRepository (with all of it's
shortcomings) except that it skipped everything that didn't match the
port/version parameters. Remove it completely and port over the explicit
port version selection into an optional parameter of _updateRepository.

Main: Factor out a _setupForPossiblyObsoletePort method and use it from
_buildMainPort and _buildPort. This ensures that the package redirection
happens in all cases and make the warning message reflect that
redirection.

This has the nice side effect that built dependencies that are actually
the latest version will not be put into the obsolete directory anymore
and therefore fixes the initial part of #22.

Also only call supportBackwardsCompatibility if the version actually
isn't the active one, since the operation is somewhat expensive and
produces otherwise confusing output.