Repository/Port: Force allow unstable if explicitly requested.
When an unstable port is explicitly requested, a prompt allows to
continue. Since the port is unstable, the actual generation of the
dependency info was still skipped however.

Force allowing unstable ports when an explicit version was requested.
Since the port packages are filtered by their status as well, also
force reparsing (and therefore recreating the package list) of the
port recipe.