Do not make the TEST phase depend on REQUIRES. (#160)
Before the introduction of the TEST_REQUIRES variable, the TEST
phase depended only on BUILD_REQUIRES. After TEST_REQUIRES was
introduced, the TEST phase started to _also_ depend on REQUIRES
and TEST_REQUIRES. The downside of making the TEST phase depend
on REQUIRES is that, in some cases, haikuporter cannot satisfy
such a dependency. That's because haikuporter considers the
REQUIRES of the base package and _every_ subpackage. This means
that if any subpackage has a dependency that cannot be met then
"--test" becomes useless for that package.

The fix for this issue is easy: make the TEST phase depend _only_
on BUILD_REQUIRES and TEST_REQUIRES.