Fix extraction of #noarchive sources
The code assumed that the directory already existed, but it doesn't
anymore since the changes to detect a mis-set SOURCE_DIR. As a result
the file would be copied in the place where a source directory was
expected.

Solution: just create the dir and then copy the file inside it.