Add a --check-ports-releases option to check for new upstream releases
This should help monitoring releases from upstream projects,
and make sure to test and update the recipes accordingly.

For now it only checks releases/tags at GitHub, maybe we should check
at repology.org instead.

Also it doesn't attempt to sort the tags, but only checks for the first one,
so in some cases it doesn't work (like with cronie-crond/cronie).

At least it works for qtkeychains which are a bit reluctant to merge Haiku
support in fear it might bitrot in the future.