python3: errors from mutating collections while iterating
This function has two loops which iterate over the items in a dictionary and
then conditionally remove from that dictionary. dict.items() returns an
iterator and not a copy of the (key, value) pairs it contains, so it isn't
safe to remove from this loop.

This makes a copy of the items by converting the iterator to a tuple.

Fixes #199.