Add a config file option to disable source packages.
Most users will probably want this always enabled (it's even very
popular to use hacks with bash aliases to do it). So it should be a
config option rather than a command line one.

However, some other command line options automatically enabe it, and so
it's hard to completely remove the command line option.

So now we have both. The source package is not built if either the
option or the configuration asks to disable it.