extractPatchset: add git config core.abbrev=auto to format-patch call
This makes sure that the patch output is consistent, even if users use
different values for that option.

`haikuporter -e` is called by developers and its output ends up in the
haikuports repository, so it should be as consistent as possible to avoid
unnecessary noise.

Other commands either don't produce any visible output or are run on the
buildmaster systems where there is a defined environment, so it isn't necessary
to define the git config explicitly.

If there are other options which influence the output of `format-patch`, they
should also be set explicitly.