Don't sign git tags
Rationale: when the git config setting `tag.gpgSign=true` is used, this implies the `-s` option to
`git tag`. This means that git needs a message for the tag. Because no message is given, git
invokes an editor, which doesn't work when running non-interactively.

We don't need signed tags in a git repo created by HaikuPorter, so the easiest solution is to pass
`--no-sign` to `git tag` calls.