buildmaster: Drop sources and provide licenses in the image.
Since building the host tools as part of the container image build, the
source volume is not actually needed for the backend and the frontend
never needed that volume in the first place.

Originally the shared source volume was meant to reduce used disk space
when running multiple instances. This is not needed anymore as the image
contains and shares the host tools and the bootstrap process for getting
the system-packages has been externalized.

The only user of the shared sources was the built in licenses in the
Haiku repository. For now, provide these in the image as well. This
could later also be moved to an external archive like what is done for
the system-packages.