Implement S3 based storage backend for PackageRepository.
The storage backend is used to hold the actual packages while the local
packages directory is only used to keep track of the current package
list.

New packages are spooled to the local packages directory as they are
built and are kept there for adding them to the package repo file
(where package information is needed and the checksum is calculated).
Once added to the repo, the packages are uploaded to object storage and
the local copy is stubbed out to an 0 byte file.

When dependency packages are needed on the builder (and are not already
cached there), they are streamed directly from object storage without
repopulating the local packages directory.

After the package repo is updated it is uploaded to object storage as
well, along with its info file, sha256 checksum and the package list
file. This allows the object storage to be used as a complete package
repo by pkgman directly.

Finally packages in object storage are then pruned based on the list of
current local stub package files to keep the state in sync.

Note that this requires a "package_repo" command that supports the "-t"
argument to the "update" command as only stub packages are available
locally and the package info can therefore not be extracted from them.
Instead the package names are assumed to be canonical and the package
info to be immutable. This is unproblematic, as the buildmaster setup
ensures that packages cannot be overwritten (this would also have failed
previously as the checksums were intentionally not revalidated).

The storage backend config file path is given with a new
"--storage-backend-config" option. It should point to a JSON file with
a "backend_type" string (only "s3" is supported for now). A sample
config is also included. An empty path is allowed and causes no storage
backend to be used.

The S3 storage backend needs an "endpoint_url", "access_key_id",
"secret_access_key" and "bucket_name" to be specified in the config
file. An optional "prefix" can also be supplied to place multiple
instances into the same bucket.

Include the storage backend config option in the buildmaster scripts fed
from a "STORAGE_BACKEND_CONFIG" environment variable for easy
configuration.