buildmaster: Drop shared packages volume.
The packages repository never actually needed to be shared or separate
and can just as well be located on the main buildmaster volume. It was
originally shared only so that repositories for multiple architectures
could be served from a single server.

When using object storage as the storage backend, the repository
directories are only used to keep the state and don't provide the
actual repo or package files. In this case a separate volume is even
less useful.

Point frontend container to the single buildmaster volume instead of the
previously shared instances directory on the packages volume. This means
that the fontend will generally not be shared across architectures
anymore. Since it reduces the scope of the shared volumes this does ease
deployment.

The "repo_consistency.txt" and "report.txt", that report the consistency
of the recipe and package repository respectively, are moved from the
packages volume to the output directory as this makes them accessible
through the normal frontend.