buildmaster: Fix debug_server config in builderctl provision.
Use SFTP to create the directory and write the file as that actually
writes the file and is less fragile.

The settings file was named "config" instead of "settings" so wouldn't
have applied anyway and the content written would just have applied for
a single "DumpRenderTree" crash. Instead set the "default_action" to
"report".