fix "any" arch subpackages on secondary arch
When determining the architecture of a package, check whether it is "any"
before checking the secondary architecture.

To make "any" packages identical to the case when not using a secondary
architecture, don't include the secondaryArchSuffix in the package name in this
case. This is a breaking change! Any packages using this must update the
PROVIDES of the subpackage and should add a REPLACES entry to facilitate
upgrades.

Add support for this change to ShellScriptlets getPackagePrefix, which is also
used by packageEntries.

Fixes #320.