fix "any" arch ports being considered buildable on secondaryArch (#335)
This fixes a regression from e1d6715aea62d5b5c0e00401a863ee197eec2258, which
erroneously considered 'any' arch ports (where the main package is "any") as
buildable on secondary architectures. Besides that this doesn't make sense
logically, in practice these packages are never buildable because the value of
$prefix is wrong in this case because the name of the main package doesn't
match the port name (and $prefix is used by all recipes directly or indirectly
because it is the target for all files to be installed to). The above commit
only added support for getPackagePrefix for the subpackages, not the main one.

Additionally, this bug caused the package-to-port associations in the
repository to become wrong and always point to the (last) secondaryArch port
instead of the primary one because secondary architectures are processed after
the primary architecture.

This change fixes the bug by marking ports with an "any" arch main package as
explicitly unsupported. This can't be done at the package level, because it
doesn't know whether it is the main package or a subpackage.