don't include secondaryArchSuffix in docDir
Documentation is usually architecture-independent, and there are multiple
packages in haikuports which in fact use an "any" arch subpackage for the
documentation. These packages then end up being different if the
$secondaryArchSuffix is included in the documentation directory name.

The alternative would be that these recipes can't use $docDir, and by extension
runConfigure and $cmakeDirArgs without overriding the doc dir, if they install
anything in it.

After this change, packages that are built for both primary and secondary
architectures should only include the documentation in only one of them to
avoid file conflicts (although I think most of them already do, and on the
other hand, these conflicts are probably not a problem if the contents are the
same).