Repository: add a function to purge stale ports (#343)
i.e. remove the work directories and downloads of old port (versions)
which have been removed. A new command line option "--purge-stale-ports"
is introduced for this.

To get the sources of ports (and thus which downloads to keep) it must
parse all recipes.

Note that this may be a bit destructive if you are switching git branches
while working on something!