defineDebugInfoPackage: change first argument to basePackageSuffix
This allows defining multiple debuginfo packages by not hard-coding the used
packageSuffix for the debuginfo package any more but deriving it from the
basePackageSuffix instead.

Note: This is a breaking change! All invocations of defineDebugInfoPackage in
haikuports must be changed. The default case to define the main debuginfo
package is to use an empty string (i.e. "") as the first argument now.
To define a debuginfo package for a subpackage, use its package suffix instead
(similar to packageEntries and getPackagePrefix).

Also adds support for PACKAGE_NAME and PACKAGE_VERSION overrides, both for the
referenced base package and the debuginfo package itself. In both cases, they
must be declared before defineDebugInfoPackage is called.

Fixes #348.

Add a missing line continuation to the usage error message.